Changed datatype of price from float to double

+++ Uebung2/lab2/src/main/java/at/ac/tuwien/big/we16/ue2/model/Bid.java
@@ -4,17 +4,17 @@ package at.ac.tuwien.big.we16.ue2.model;
  * Created by mstrasser on 4/12/16.
  */
 public class Bid {
-    public Bid(User u, float price) {
+    public Bid(User u, double price) {
         this.user = u;
         this.price = price;
     }
 
     private Product product;
     private User user;
-    private float price;
+    private double price;
 
     public User getUser() { return this.user; }
-    public float getPrice() { return this.price; }
+    public double getPrice() { return this.price; }
 
     public Product getProduct() {
         return product;
